+++ tapo-py/src/handlers/child_devices.rs
@@ -1,3 +1,4 @@
+mod ke100_handler;
 mod power_strip_plug_handler;
 mod s200b_handler;
 mod t100_handler;
@@ -5,6 +6,7 @@ mod t110_handler;
 mod t300_handler;
 mod t31x_handler;
 
+pub use ke100_handler::*;
 pub use power_strip_plug_handler::*;
 pub use s200b_handler::*;
 pub use t100_handler::*;