+++ tapo/src/api/child_devices/ke100_handler.rs
@@ -4,8 +4,9 @@ use tokio::sync::RwLock;
 
 use crate::api::ApiClient;
 use crate::error::{Error, TapoResponseError};
+use crate::requests::TemperatureUnitKE100;
 use crate::requests::{EmptyParams, TapoParams, TapoRequest, TrvSetDeviceInfoParams};
-use crate::responses::{DecodableResultExt, KE100Result, TemperatureUnitKE100};
+use crate::responses::{DecodableResultExt, KE100Result};
 
 /// Handler for the [KE100](https://www.tp-link.com/en/search/?q=KE100) devices.
 pub struct KE100Handler {
@@ -45,87 +46,13 @@ impl KE100Handler {
             .ok_or_else(|| Error::Tapo(TapoResponseError::EmptyResult))
     }
 
-    /// Sets the *target temperature*.
-    ///
+    /// Sets *child protection* on the device to *on* or *off*.
+    ///     
     /// # Arguments
     ///
-    /// * `target_temperature` - between `min_control_temperature` and `max_control_temperature`
-    /// * `temperature_unit`
-    pub async fn set_target_temperature(
-        &self,
-        target_temperature: u8,
-        temperature_unit: TemperatureUnitKE100,
-    ) -> Result<(), Error> {
-        let device_info = self.get_device_info().await?;
-
-        if target_temperature < device_info.min_control_temperature
-            || target_temperature > device_info.max_control_temperature
-        {
-            return Err(Error::Validation {
-                field: "target_temperature".to_string(),
-                message: format!(
-                    "Target temperature must be between {} (min_control_temperature) and {} (max_control_temperature)",
-                    device_info.min_control_temperature, device_info.max_control_temperature
-                ),
-            });
-        }
-
-        let json = serde_json::to_value(
-            TrvSetDeviceInfoParams::new()
-                .target_temperature(target_temperature, temperature_unit)?,
-        )?;
-        let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
-
-        self.client
-            .read()
-            .await
-            .control_child::<serde_json::Value>(self.device_id.clone(), request)
-            .await?;
-
-        Ok(())
-    }
-
-    /// Sets the *minimal control temperature*.
-    ///
-    /// # Arguments
-    ///
-    /// * `min_control_temperature`
-    /// * `temperature_unit`
-    pub async fn set_min_control_temperature(
-        &self,
-        min_control_temperature: u8,
-        temperature_unit: TemperatureUnitKE100,
-    ) -> Result<(), Error> {
-        let json = serde_json::to_value(
-            TrvSetDeviceInfoParams::new()
-                .min_control_temperature(min_control_temperature, temperature_unit)?,
-        )?;
-        let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
-
-        self.client
-            .read()
-            .await
-            .control_child::<serde_json::Value>(self.device_id.clone(), request)
-            .await?;
-
-        Ok(())
-    }
-
-    /// Sets the *maximum control temperature*.
-    ///
-    /// # Arguments
-    ///
-    /// * `max_control_temperature`
-    /// * `temperature_unit`
-    pub async fn set_max_control_temperature(
-        &self,
-        max_control_temperature: u8,
-        temperature_unit: TemperatureUnitKE100,
-    ) -> Result<(), Error> {
-        let json = serde_json::to_value(
-            TrvSetDeviceInfoParams::new()
-                .max_control_temperature(max_control_temperature, temperature_unit)?,
-        )?;
+    /// * `on`
+    pub async fn set_child_protection(&self, on: bool) -> Result<(), Error> {
+        let json = serde_json::to_value(TrvSetDeviceInfoParams::new().child_protection(on)?)?;
         let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
 
         self.client
@@ -141,10 +68,33 @@ impl KE100Handler {
     ///     
     /// # Arguments
     ///
-    /// * `frost_protection_on` - true/false
-    pub async fn set_frost_protection(&self, frost_protection_on: bool) -> Result<(), Error> {
+    /// * `on`
+    pub async fn set_frost_protection(&self, on: bool) -> Result<(), Error> {
+        let json = serde_json::to_value(TrvSetDeviceInfoParams::new().frost_protection_on(on)?)?;
+        let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
+
+        self.client
+            .read()
+            .await
+            .control_child::<serde_json::Value>(self.device_id.clone(), request)
+            .await?;
+
+        Ok(())
+    }
+
+    /// Sets the *maximum control temperature*.
+    ///
+    /// # Arguments
+    ///
+    /// * `value`
+    /// * `unit`
+    pub async fn set_max_control_temperature(
+        &self,
+        value: u8,
+        unit: TemperatureUnitKE100,
+    ) -> Result<(), Error> {
         let json = serde_json::to_value(
-            TrvSetDeviceInfoParams::new().frost_protection_on(frost_protection_on)?,
+            TrvSetDeviceInfoParams::new().max_control_temperature(value, unit)?,
         )?;
         let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
 
@@ -157,15 +107,58 @@ impl KE100Handler {
         Ok(())
     }
 
-    /// Sets *child protection* on the device to *on* or *off*.
-    ///     
+    /// Sets the *minimum control temperature*.
+    ///
     /// # Arguments
     ///
-    /// * `child_protection_on` - true/false
-    pub async fn set_child_protection(&self, child_protection_on: bool) -> Result<(), Error> {
+    /// * `value`
+    /// * `unit`
+    pub async fn set_min_control_temperature(
+        &self,
+        value: u8,
+        unit: TemperatureUnitKE100,
+    ) -> Result<(), Error> {
         let json = serde_json::to_value(
-            TrvSetDeviceInfoParams::new().child_protection(child_protection_on)?,
+            TrvSetDeviceInfoParams::new().min_control_temperature(value, unit)?,
         )?;
+        let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
+
+        self.client
+            .read()
+            .await
+            .control_child::<serde_json::Value>(self.device_id.clone(), request)
+            .await?;
+
+        Ok(())
+    }
+
+    /// Sets the *target temperature*.
+    ///
+    /// # Arguments
+    ///
+    /// * `value` - between `min_control_temperature` and `max_control_temperature`
+    /// * `unit`
+    pub async fn set_target_temperature(
+        &self,
+        value: u8,
+        unit: TemperatureUnitKE100,
+    ) -> Result<(), Error> {
+        let device_info = self.get_device_info().await?;
+
+        if value < device_info.min_control_temperature
+            || value > device_info.max_control_temperature
+        {
+            return Err(Error::Validation {
+                field: "target_temperature".to_string(),
+                message: format!(
+                    "Target temperature must be between {} (min_control_temperature) and {} (max_control_temperature)",
+                    device_info.min_control_temperature, device_info.max_control_temperature
+                ),
+            });
+        }
+
+        let json =
+            serde_json::to_value(TrvSetDeviceInfoParams::new().target_temperature(value, unit)?)?;
         let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
 
         self.client
@@ -181,17 +174,15 @@ impl KE100Handler {
     ///
     /// # Arguments
     ///
-    /// * `temperature_offset` - between -10 and 10
-    /// * `temperature_unit`
+    /// * `value` - between -10 and 10
+    /// * `unit`
     pub async fn set_temperature_offset(
         &self,
-        temperature_offset: i8,
-        temperature_unit: TemperatureUnitKE100,
+        value: i8,
+        unit: TemperatureUnitKE100,
     ) -> Result<(), Error> {
-        let json = serde_json::to_value(
-            TrvSetDeviceInfoParams::new()
-                .temperature_offset(temperature_offset, temperature_unit)?,
-        )?;
+        let json =
+            serde_json::to_value(TrvSetDeviceInfoParams::new().temperature_offset(value, unit)?)?;
         let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
 
         self.client